requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
langchain
langchain-openai
langchain-chroma
cohere
pandas
PyYAML
openpyxl
deep-translator
langdetect
protobuf==4.21.12
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
tiktoken

